Test packages separated.
License comment cleanup.
