+++ app/src/androidTest/java/com/alexvt/integrity/test/data/RoomMetadataRepositoryTest.kt
@@ -4,7 +4,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package com.alexvt.integrity
+package com.alexvt.integrity.test.data
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
