+++ app/src/test/java/com/alexvt/integrity/test/ui/MainScreenViewModelTest.kt
@@ -4,7 +4,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package com.alexvt.integrity
+package com.alexvt.integrity.test.ui
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.lifecycle.Lifecycle
