+++ integrity_lib/src/main/java/com/alexvt/integrity/lib/android/util/TypeSpecificMetadataConverter.kt
@@ -4,12 +4,6 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-/*
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-
 package com.alexvt.integrity.lib.android.util
 
 import android.content.Context
